+++ blacksheep/server/controllers.py
@@ -1,8 +1,9 @@
 import sys
 from io import BytesIO
-from typing import Any, AsyncIterable, Callable, Optional, Union
+from typing import Any, AsyncIterable, Callable, ClassVar, List, Optional, Type, Union
 
 from blacksheep import Request, Response
+from blacksheep.common.types import HeadersType, ParamsType
 from blacksheep.server.responses import (
     ContentDispositionType,
     MessageType,
@@ -28,7 +29,7 @@ from blacksheep.server.responses import (
     view,
     view_async,
 )
-from blacksheep.server.routing import RoutesRegistry
+from blacksheep.server.routing import ActionFilter, RoutesRegistry, normalize_filters
 from blacksheep.utils import AnyStr, join_fragments
 
 # singleton router used to store initial configuration,
@@ -48,6 +49,35 @@ trace = router.trace
 options = router.options
 connect = router.connect
 ws = router.ws
+
+
+def filters(
+    *filters: ActionFilter,
+    host: Optional[str] = None,
+    headers: Optional[HeadersType] = None,
+    params: Optional[ParamsType] = None,
+):
+    """
+    Configures a set of filters for a decorated controller type.
+    Filters are applied to all routes defined in the controller.
+
+    ---
+    Example: match a "/" route only if the request includes an header "X-Area: Special"
+
+    ```py
+    @filters(headers={"X-Area": "Special"})
+    class Special(Controller):
+        @get("/")
+        def special(self):
+            ...
+    ```
+    """
+
+    def class_deco(cls: Type["Controller"]):
+        cls._filters_ = normalize_filters(host, headers, params, list(*filters))
+        return cls
+
+    return class_deco
 
 
 class CannotDetermineDefaultViewNameError(RuntimeError):
@@ -71,6 +101,8 @@ class ControllerMeta(type):
 
 class Controller(metaclass=ControllerMeta):
     """Base class for all controllers."""
+
+    _filters_: ClassVar[List[ActionFilter]] = []
 
     @classmethod
     def route(cls) -> Optional[str]: